+++ RELEASE_NOTES.md
@@ -1,12 +1,16 @@
 ## Release Notes 3.0.6
 
+- **VPin Mania**: Added VPS table search.
 - **Preferences / Notifications**: Added "Desktop Mode" checkbox for notifications.
 - **Table Overview / Table Data Manager**: Optimized dialog so that it can be displayed for smaller screen resolutions.
+- **iScored Integration**: Added info button which shows a summary of your game room settings.
+
+  <img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/mania/gameroom-info.png" width="600" />
+
 
 **Bugfixes**
 
 - **VPin Mania**: The actual highscore timestamp is now used when a highscore is written. So when you synchronize your highscores, the actual creation date will be used instead of the date of submission. This way, the "Recent Highscores" widget shows only the actual recently created highscores.
 - **VPin Mania**: Fixed a bunch of sorting and navigation issues.
-- **VPin Mania**: Added VPS table search.
 - **VPS Table Search Input**: Fixed input fields for VPS table searches so that whitespaces are allowed now too.
 - **Mac OS**: Fixed wrong OS name detection when editing files.